# aphrodite-backend/docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "aphrodite",
        "description": "API Documentation",
        "version": "unknown"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8083/aphrodite-web-api",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/key-value/data/{bucket}": {
            "get": {
                "tags": ["KeyValueData"],
                "operationId": "readKeyValueData",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Secret Key Version",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        },
                        "example": 1
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "put": {
                "tags": ["KeyValueData"],
                "operationId": "writeKeyValueData",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Secret Key Version",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        },
                        "example": 1
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WriteKeyValueDataResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "delete": {
                "tags": ["KeyValueData"],
                "operationId": "deleteKeyValueData",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Secret Key Version",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        },
                        "example": 1
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
        "/api/v1/buckets": {
            "get": {
                "tags": ["Buckets"],
                "summary": "List all buckets in the current namespace",
                "operationId": "listBuckets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "post": {
                "tags": ["Buckets"],
                "summary": "Create a bucket in the current namespace",
                "operationId": "createBucket",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BucketRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bucket"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
        "/api/v1/buckets/{bucketName}": {
            "delete": {
                "tags": ["Buckets"],
                "summary": "Delete an existing bucket name in the current namespace",
                "operationId": "deleteBucket",
                "parameters": [
                    {
                        "name": "bucketName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        }
    },
    "components": {
        "schemas": {
...
            "WriteKeyValueDataResponse": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "BucketRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    }
                }
            },
            "Bucket": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Aphrodite-API-Key": {
                "type": "apiKey",
                "name": "Aphrodite-API-Key",
                "in": "header"
            }
        }
    }
}